+++ Source/AvFilterFactoryBase.h
@@ -1,8 +1,8 @@
 #pragma once
 #include "pch.h"
-#include "IAvEffect.h"
+#include "IAvFilter.h"
 
-class AbstractEffectFactory
+class AvFilterFactoryBase
 {
 public:
     virtual std::shared_ptr<IAvFilter> CreateEffect(winrt::hstring const& definitions) = 0;
